requirements.txt
Django==3.0.4
Pillow==7.0.0
django-cleanup==4.0.0
django-debug-toolbar==2.2
